feat: Complete React anti-patterns fixes (Issue #7)
✅ Missing keys: All React keys properly implemented (already fixed)
✅ Inline styles: Replaced with CSS modules for better maintainability
✅ TypeScript types: Proper type definitions with shared interfaces
✅ Poor component structure: Broke monolithic component into focused components

Component architecture improvements:
- Created modular components: AdvocateRow, LoadingSpinner, ErrorMessage, FilterControls, AdvocatesTable, NoResults
- Added shared types: Advocate interface in dedicated types file
- Implemented CSS modules: advocates.module.css with organized styles
- Added TypeScript declarations: CSS modules support

Technical improvements:
- Eliminated 38+ inline style instances
- Separated concerns: Each component has single responsibility
- Improved maintainability: Styles centralized in CSS modules
- Better reusability: Components can be easily reused or tested
- Enhanced type safety: Shared Advocate interface across components

File structure:
- src/app/components/ - Modular React components
- src/app/types/ - Shared TypeScript interfaces
- src/app/advocates.module.css - Centralized styles
- src/app/page.tsx - Clean main component orchestrating sub-components

All React anti-patterns resolved with modern best practices!
